requirements.txt
fastapi
uvicorn==0.23.0
gunicorn==20.1.0
uvloop
httptools
sqlalchemy
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
alembic
psycopg2-binary
python-decouple
python-multipart
email-validator
python-jose[cryptography]
httpx
faker
pytest
pydantic
python-dotenv
requests
email_validator